feat: Update UserItems.tsx and Mock.ts to add new user items and remove unused code
